feat: inicia expense

+++ src/user/user.service.ts
@@ -1,6 +1,10 @@
-import { Injectable, NotFoundException } from '@nestjs/common';
+import {
+  ConflictException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { UserRepository } from './user.repository';
-import { UserEntity } from './user.entity';
+import { User } from './user.entity';
 import { UpdateResult } from 'typeorm';
 import { CreateUserDto } from './create-user.dto';
 import { UpdateUserDto } from './update-user.dto';
@@ -8,15 +12,15 @@ import { UpdateUserDto } from './update-user.dto';
 @Injectable()
 export class UserService {
   constructor(private readonly userRepository: UserRepository) {}
-  async create(userDto: CreateUserDto): Promise<UserEntity> {
+  async create(userDto: CreateUserDto): Promise<User> {
     return this.userRepository.save(userDto);
   }
 
-  async getAll(): Promise<UserEntity[]> {
+  async getAll(): Promise<User[]> {
     return this.userRepository.find();
   }
 
-  async getOne(id: string): Promise<UserEntity> {
+  async getOne(id: string): Promise<User> {
     const user = await this.userRepository.findOne(id);
     if (!user) {
       throw new NotFoundException('Usuário não encontrado');
@@ -24,7 +28,7 @@ export class UserService {
     return user;
   }
 
-  async getByEmail(email: string): Promise<UserEntity> {
+  async getByEmail(email: string): Promise<User> {
     const user = await this.userRepository.findByEmail(email);
     if (!user) {
       throw new NotFoundException('Usuário não encontrado');
@@ -40,10 +44,13 @@ export class UserService {
     return this.userRepository.update(user.id, userDto);
   }
 
-  async deleteById(id: string): Promise<UserEntity> {
+  async deleteById(id: string): Promise<User> {
     const user = await this.userRepository.findOne(id);
     if (!user) {
       throw new NotFoundException('Usuário não encontrado');
+    }
+    if (user.expenses.length > 0) {
+      throw new ConflictException('Usuário com despesas não pode ser removido');
     }
     return this.userRepository.remove(user);
   }
